+++ include/modules/colors/renderers/circle_animation_renderer.hpp
@@ -11,7 +11,7 @@ public:
   CircleAnimationRenderer(BetterCanvas *canvas, bool _shrink = false);
   virtual ~CircleAnimationRenderer();
 
-  virtual int render() override;
+  virtual long int render() override;
   virtual void setup() override;
   virtual void teardown() override;
